backend/requirements.txt
fastapi[all]
# redis
# types-redis
uvicorn
numpy


pydantic
python-multipart

# Database
sqlalchemy
psycopg2-binary

# Authentication
python-jose[cryptography]
passlib[bcrypt]
pyjwt
Authlib==1.3.2

# Additional utilities
python-dotenv
email-validator

loguru
boto3


fastapi-mail 
python-multipart 
jinja2

# For Google Auth
google-auth==2.35.0
websockets==13.1
cryptography==43.0.1

httpx==0.27.2
urllib3==2.2.3
requests

# for session
itsdangerous
starlette>=0.27.0

